requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx==0.24.1
beautifulsoup4==4.9.3
python-multipart==0.0.5
fake-useragent==0.1.11
aiohttp==3.8.1
python-dotenv==0.19.0
redis==4.0.2
python-jose[cryptography]==3.3.0
lxml==4.9.1
markupsafe==2.0.1
pydantic==1.8.2